refactor: extract setup feature, profile scope, and plan services
