+++ src/Func/Hosting/BuiltInCommands.cs
@@ -47,6 +47,9 @@ internal static class BuiltInCommands
         services.AddSingleton<FuncCliCommand, ProfileCommand>();
 
         // Setup command and dependencies
+        services.AddSingleton<ISetupFeatureResolver, SetupFeatureResolver>();
+        services.AddSingleton<ISetupProfileScopeResolver, SetupProfileScopeResolver>();
+        services.AddSingleton<ISetupDependencyPlanBuilder, SetupDependencyPlanBuilder>();
         services.AddSingleton<ISetupDependencyInstaller, SetupDependencyInstaller>();
         services.AddSingleton<ISetupRunner, SetupRunner>();
         services.AddSingleton<FuncCliCommand, SetupCommand>();
