+++ test/Func.Tests/Commands/Setup/SetupRunnerTests.cs
@@ -730,13 +730,13 @@ public sealed class SetupRunnerTests : IDisposable
         InteractiveTestInteractionService interactive = new();
         SetupRunner runner = new(
             interactive,
-            _store,
-            new SetupDependencyInstaller(interactive, _store, catalog, _installer),
-            _profileCatalog,
-            new TestOptionsMonitor<ProjectProfileOptions>(new ProjectProfileOptions()),
-            new TestOptionsMonitor<UserProfilePreferenceOptions>(new UserProfilePreferenceOptions()),
-            new FakeCliConfigurationProvider(new Dictionary<string, string?>()),
-            _bundleReader);
+            new SetupFeatureResolver(interactive, _store, new FakeCliConfigurationProvider(new Dictionary<string, string?>())),
+            new SetupProfileScopeResolver(
+                _profileCatalog,
+                new TestOptionsMonitor<ProjectProfileOptions>(new ProjectProfileOptions()),
+                new TestOptionsMonitor<UserProfilePreferenceOptions>(new UserProfilePreferenceOptions())),
+            new SetupDependencyPlanBuilder(_bundleReader),
+            new SetupDependencyInstaller(interactive, _store, catalog, _installer));
 
         SetupRunResult result = await runner.RunAsync(
             new SetupCommandOptions(
@@ -803,13 +803,13 @@ public sealed class SetupRunnerTests : IDisposable
         InteractiveTestInteractionService interactive = new();
         SetupRunner runner = new(
             interactive,
-            _store,
-            new SetupDependencyInstaller(interactive, _store, catalog, _installer),
-            _profileCatalog,
-            new TestOptionsMonitor<ProjectProfileOptions>(new ProjectProfileOptions()),
-            new TestOptionsMonitor<UserProfilePreferenceOptions>(new UserProfilePreferenceOptions()),
-            new FakeCliConfigurationProvider(new Dictionary<string, string?>()),
-            _bundleReader);
+            new SetupFeatureResolver(interactive, _store, new FakeCliConfigurationProvider(new Dictionary<string, string?>())),
+            new SetupProfileScopeResolver(
+                _profileCatalog,
+                new TestOptionsMonitor<ProjectProfileOptions>(new ProjectProfileOptions()),
+                new TestOptionsMonitor<UserProfilePreferenceOptions>(new UserProfilePreferenceOptions())),
+            new SetupDependencyPlanBuilder(_bundleReader),
+            new SetupDependencyInstaller(interactive, _store, catalog, _installer));
 
         _ = await runner.RunAsync(
             new SetupCommandOptions(
@@ -844,13 +844,13 @@ public sealed class SetupRunnerTests : IDisposable
             .WithLatest(IInstalledBundleWorkloads.BundleWorkloadPackageId, "4.10.0");
         SetupRunner runner = new(
             _interaction,
-            _store,
+            new SetupFeatureResolver(_interaction, _store, new FakeCliConfigurationProvider(new Dictionary<string, string?>())),
+            new SetupProfileScopeResolver(
+                _profileCatalog,
+                new TestOptionsMonitor<ProjectProfileOptions>(new ProjectProfileOptions()),
+                new TestOptionsMonitor<UserProfilePreferenceOptions>(new UserProfilePreferenceOptions())),
+            new SetupDependencyPlanBuilder(_bundleReader),
             new SetupDependencyInstaller(_interaction, _store, catalog, _installer),
-            _profileCatalog,
-            new TestOptionsMonitor<ProjectProfileOptions>(new ProjectProfileOptions()),
-            new TestOptionsMonitor<UserProfilePreferenceOptions>(new UserProfilePreferenceOptions()),
-            new FakeCliConfigurationProvider(new Dictionary<string, string?>()),
-            _bundleReader,
             firstRunStore);
 
         SetupRunResult result = await runner.RunAsync(Options(), CancellationToken.None);
@@ -865,13 +865,13 @@ public sealed class SetupRunnerTests : IDisposable
         IProfileCatalog? profileCatalog = null)
         => new(
             _interaction,
-            _store,
-            new SetupDependencyInstaller(_interaction, _store, catalog, _installer),
-            profileCatalog ?? _profileCatalog,
-            new TestOptionsMonitor<ProjectProfileOptions>(new ProjectProfileOptions()),
-            new TestOptionsMonitor<UserProfilePreferenceOptions>(new UserProfilePreferenceOptions()),
-            new FakeCliConfigurationProvider(projectConfig ?? new Dictionary<string, string?>()),
-            _bundleReader);
+            new SetupFeatureResolver(_interaction, _store, new FakeCliConfigurationProvider(projectConfig ?? new Dictionary<string, string?>())),
+            new SetupProfileScopeResolver(
+                profileCatalog ?? _profileCatalog,
+                new TestOptionsMonitor<ProjectProfileOptions>(new ProjectProfileOptions()),
+                new TestOptionsMonitor<UserProfilePreferenceOptions>(new UserProfilePreferenceOptions())),
+            new SetupDependencyPlanBuilder(_bundleReader),
+            new SetupDependencyInstaller(_interaction, _store, catalog, _installer));
 
     private SetupCommandOptions Options(
         IReadOnlyList<string>? features = null,
